refactor: versão sem o produtor-consumidor

+++ trabalho03/src/main/java/banheirounissex/Person.java
@@ -3,13 +3,13 @@ package banheirounissex;
 public class Person {
     private Integer id;
     private char gender;
-    private Integer time;
+    private Long time;
 
     public Person() {
 
     }
 
-    public Person(Integer id, char gender, Integer time) {
+    public Person(Integer id, char gender, Long time) {
         this.id = id;
         this.gender = gender;
         this.time = time;
@@ -31,11 +31,11 @@ public class Person {
         this.gender = gender;
     }
 
-    public Integer getTime() {
+    public Long getTime() {
         return this.time;
     }
 
-    public void setTime(Integer time) {
+    public void setTime(Long time) {
         this.time = time;
     }
